Made ReadAllAccountRequest take a optional name query parameter. (#97)
Also introduced QueryParamBuilder and TwilioStringValue abstraction.

And added TwilioStringValue trait, as an abstraction for all classes that reprent a value that can be used in the Twilio API as a String. This abstraction is used by QueryParamBuilder. It is not added to all classes that should have it, but only the onces that was needed for now.