CHA-957 Created SidAbstract as a base class for all Sid classes.
Also added an abstract class for hte companion objects, making is super easy to create a seperate set of error classes, and corosponding factory methods.

Also create a base type for tests of such clases, making is super easy to create a full test suite for a Sid class.

Made Call.Sid make use of this new class.

This is a breaking change, as I change apply method from being a unsafe factory, to return an either.