CHA-990 Refactored call a bit, so it now uses java.time.Duration for duration and queueTime, instead of its own String wrapper class.
Makes it a lot easier to handle for clients, as they then don't need to now what the string represent in the different cases. That part was actually confusing, because it is a string in the first place, and in some case the string is an integer representing seconds, and in the other case is representing milliseconds.