CHA-1014 Refactored the wait part out of DtmfDigit and into DtmfString.
wait (w) is not a real DTMF digit, but it is often used in Strings of DTMF. So I have refactored it out of DtmfDigit, so that enum know only represents real DTMF digits. Instead, I have refactored DtmfString a lot, so it now has two subtypes, one for representing pure strings of DTMF digits, and one that also support having waits in the string. This better reflect reality, and makes it possible to reuse the DtmfDigit enum in places where it requires only real DTMF digits as input, such as the upcoming implementation of the Gather directive would need.